# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend normativo de acceso a derechos",
    "description": "\nAPI de consulta del corpus normativo de acceso a derechos.\n\nToda respuesta lleva la misma envoltura: `schema_version`, `release_id`,\n`as_of`, `known_at`, `data_status`, `evidence`, `missing_fields` y `warnings`.\nSin `as_of` una respuesta no dice para cuándo vale; sin `release_id` no se puede\nreproducir; sin `missing_fields` una abstención sería indistinguible de un \"no\".\n\nLos errores son tipados. Información insuficiente es un estado de dominio y\nnunca se disfraza de 500.\n\nEsta API no otorga, no rechaza y no revoca prestaciones. Devuelve lo que dicen\nlas normas publicadas, con la evidencia que lo sostiene y lo que falta averiguar.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/beneficios": {
      "get": {
        "tags": [
          "operativo"
        ],
        "summary": "Listar Beneficios",
        "description": "Beneficios candidatos.\n\nPertenecer a una población no implica tener derecho: el filtro acota la\nbúsqueda, no infiere elegibilidad. Para saber si corresponde hay que evaluar\nlas reglas, que es lo que hace `POST /v1/evaluaciones-preliminares`.",
        "operationId": "listar_beneficios_v1_beneficios_get",
        "parameters": [
          {
            "name": "poblacion",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Código de población destinataria.",
              "title": "Poblacion"
            },
            "description": "Código de población destinataria."
          },
          {
            "name": "territorio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Jurisdicción.",
              "title": "Territorio"
            },
            "description": "Jurisdicción."
          },
          {
            "name": "linea",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Linea"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limite"
            }
          },
          {
            "name": "desplazamiento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desplazamiento"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_Pagina_BeneficioResumen__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/valores": {
      "get": {
        "tags": [
          "operativo"
        ],
        "summary": "Valores",
        "description": "Valor de un parámetro para la fecha consultada.\n\nNunca se devuelve el último valor por fecha máxima: se devuelve el que se\naplica a la fecha pedida. Servir el monto de otro período es la forma más\ndirecta de dar una respuesta incorrecta que parece correcta.",
        "operationId": "valores_v1_valores_get",
        "parameters": [
          {
            "name": "concepto",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Código del parámetro.",
              "title": "Concepto"
            },
            "description": "Código del parámetro."
          },
          {
            "name": "territorio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Territorio"
            }
          },
          {
            "name": "segmento",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Segmento"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_list_ValorParametro__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/plazos": {
      "get": {
        "tags": [
          "operativo"
        ],
        "summary": "Plazos",
        "description": "Plazos con su tipo y su estado calculado a la fecha.\n\nUn plazo relativo sin evento de inicio conocido responde \"desconocido\", no\nuna fecha inventada.",
        "operationId": "plazos_v1_plazos_get",
        "parameters": [
          {
            "name": "beneficio_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Beneficio Id"
            }
          },
          {
            "name": "tramite_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tramite Id"
            }
          },
          {
            "name": "convocatoria",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Convocatoria"
            }
          },
          {
            "name": "ciclo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ciclo"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_list_PlazoResumen__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/puntos-atencion": {
      "get": {
        "tags": [
          "operativo"
        ],
        "summary": "Puntos Atencion",
        "description": "Puntos de atención con sus canales.\n\nSin coordenadas válidas no se afirma cercanía. El horario es por canal: el\nde la mesa presencial no se copia al teléfono.",
        "operationId": "puntos_atencion_v1_puntos_atencion_get",
        "parameters": [
          {
            "name": "jurisdiccion",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jurisdiccion"
            }
          },
          {
            "name": "localidad",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Localidad"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_list_PuntoAtencion__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evaluaciones-preliminares": {
      "post": {
        "tags": [
          "evaluación"
        ],
        "summary": "Evaluar",
        "operationId": "evaluar_v1_evaluaciones_preliminares_post",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudEvaluacion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_ResultadoEvaluacion_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Respuesta_Pagina_BeneficioResumen__": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "release_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Id"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "known_at": {
            "type": "string",
            "format": "date-time",
            "title": "Known At"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "data": {
            "$ref": "#/components/schemas/Pagina_BeneficioResumen_"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidencia"
            },
            "type": "array",
            "title": "Evidence"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Advertencia"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "known_at",
          "data_status",
          "data"
        ],
        "title": "Respuesta[Pagina[BeneficioResumen]]"
      },
...
    }
  },
...
}